Prototype for Study Planner
This is a very, VERY early and rough prototype draft of the study planner, as part of our larger project. The HTML, JS, and CSS files are a slightly updated version of a drawn out prototype done on paper, drawn by hand (Seen in pictures). Please be patient, this is a very early stage of a work in progress, and more changes are definitely in the works. 

God bless, 

Andre Dudley Jr.